Merge From develop into frontend

See merge request s09-bigdata-dist-sub2/S09P22B302!171

+++ zizon/src/main/java/com/b302/zizon/domain/playlist/dto/PlayPlaylistResponseDTO.java
@@ -9,11 +9,11 @@ public class PlayPlaylistResponseDTO {
     private Long musicId;
     private String title;
     private String artist;
-    private int duration;
+    private Long duration;
     private String youtubeVideoId;
     private String albumCoverUrl;
 
-    public PlayPlaylistResponseDTO(Long musicId, String title, String artist, int duration, String youtubeVideoId, String albumCoverUrl) {
+    public PlayPlaylistResponseDTO(Long musicId, String title, String artist, Long duration, String youtubeVideoId, String albumCoverUrl) {
         this.musicId = musicId;
         this.title = title;
         this.artist = artist;
